// frontend/src/features/Map/components/Overlay/types.ts
export type OverlayCardMargins = {
  xLeft: number
  xMiddle: number
  xRight: number
  yBottom: number
  yMiddle: number
  yTop: number
}

export enum OverlayPosition {
  BOTTOM = 'BOTTOM',
  BOTTOM_LEFT = 'BOTTOM_LEFT',
  BOTTOM_RIGHT = 'BOTTOM_RIGHT',
  LEFT = 'LEFT',
  RIGHT = 'RIGHT',
  TOP = 'TOP',
  TOP_LEFT = 'TOP_LEFT',
  TOP_RIGHT = 'TOP_RIGHT'
}